fix: operand size issues with HostToNetworkOrder (defaulted to int without explicitely casting as short and later conversion to ushort), now solving ref id for replies and dropping responses to requests that were already resolved and fixed timeout in the join state to not reset after each message is received

+++ IPK25-CHAT/Message/ErrMessage.cs
@@ -53,7 +53,7 @@ public class ErrMessage(string displayName, string messageContent, ushort messag
         MessageID = messageID;
 
         // Message fields
-        byte[] messageIDBytes = BitConverter.GetBytes(IPAddress.HostToNetworkOrder(MessageID));
+        byte[] messageIDBytes = BitConverter.GetBytes(IPAddress.HostToNetworkOrder((short)MessageID));
         byte[] displayNameBytes = Encoding.ASCII.GetBytes(DisplayName);
         byte[] messageContentBytes = Encoding.ASCII.GetBytes(MessageContent);
